Moving BST model test
