+++ tests/tasks/test_blended_skill_talk.py
@@ -402,31 +402,5 @@ class TestBlendedSkillTalkInteractiveWorld(unittest.TestCase):
             self.assertEqual(len(personas), count)
 
 
-class TestBlendedSkillTalkModels(unittest.TestCase):
-    def test_accuracies(self):
-        """
-        Test accuracies of BST models in the zoo.
-        """
-        model_names_to_desired_accuracies = {
-            'bst_single_task': 0.8125,
-            'convai2_single_task': 0.7715,
-            'ed_single_task': 0.7148,
-            'wizard_single_task': 0.6797,
-            'multi_task': 0.8086,
-            'multi_task_bst_tuned': 0.8457,
-        }
-        for model_name, desired_accuracy in model_names_to_desired_accuracies.items():
-            valid, _ = testing_utils.eval_model(
-                opt={
-                    'batchsize': 64,
-                    'model_file': f'zoo:blended_skill_talk/{model_name}/model',
-                    'task': 'blended_skill_talk',
-                    'num_examples': 1000,
-                },
-                skip_test=True,
-            )
-            self.assertAlmostEqual(valid['accuracy'], desired_accuracy, delta=0.005)
-
-
 if __name__ == '__main__':
     unittest.main()
